pyproject.toml
[tool.poetry]
name = "mcp-server"
version = "0.1.0"
description = "Multi-Component Platform Server with authenticated tools"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
flask-cors = "^4.0.0"
flask-jwt-extended = "^4.5.0"
pandas = "^2.0.3"
openpyxl = "^3.1.2"
plotly = "^5.15.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
py-opa = "^0.1.0"
python-json-logger = "^2.0.7"
streamlit = "^1.25.0"
uvicorn = "^0.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.app:main"
dev = "src.app:main"